requirements.txt
streamlit==1.34.0
google-generativeai==0.5.4
python-dotenv==1.0.1